# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Cortex Plane Control API
  version: 1.0.0
  description: |
    REST API for the Cortex Plane autonomous agent orchestration platform.

    The control plane is a stateless orchestrator that dispatches work to
    per-agent containers on a k3s cluster.  PostgreSQL is the source of truth
    for all authoritative state; Qdrant provides semantic search over
    long-term memory.

    ## Authentication

    Three authentication mechanisms are supported, depending on the caller:

    | Mechanism | Header | Use case |
    |-----------|--------|----------|
    | API key | `X-API-Key: <key>` | External clients, CI/CD pipelines |
    | Session token | `Authorization: Bearer <session-id>` | Dashboard, agent SSE streams |
    | k8s ServiceAccount | `Authorization: Bearer <sa-token>` | In-cluster agents, init containers |

    All endpoints except `/healthz`, `/readyz`, and `/approvals/stream`
    require authentication.

    ## Rate Limiting

    Rate limits are enforced per API key / session and returned in response
    headers:

    | Header | Description |
    |--------|-------------|
    | `X-RateLimit-Limit` | Max requests in the current window |
    | `X-RateLimit-Remaining` | Requests remaining in the current window |
    | `X-RateLimit-Reset` | UTC epoch seconds when the window resets |

    **Default tiers:**

    | Tier | Limit | Applies to |
    |------|-------|------------|
    | Standard | 60 req/min | Read endpoints (`GET`) |
...

...

...

...

...

...

...
    {
...

...

...

...

...

...
servers:
...
security:
  - apiKey: []
  - bearerAuth: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Static API key for external clients and CI/CD pipelines.
        Issued by the platform operator and stored as a k8s Secret.

    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Session token (UUID) or Kubernetes ServiceAccount JWT.
        Session tokens are validated against the `session` table.
        ServiceAccount tokens are validated via the TokenReview API.

  # -----------------------------------------------------------------------
  # Headers
  # -----------------------------------------------------------------------
  headers:
    X-RateLimit-Limit:
      description: Maximum requests allowed in the current window.
      schema:
        type: integer
        example: 60
...
  schemas:
...
    Pagination:
      type: object
      required:
        - total
        - limit
        - offset
        - hasMore
      properties:
        total:
          type: integer
          minimum: 0
          example: 42
        limit:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        offset:
          type: integer
          minimum: 0
          example: 0
        hasMore:
          type: boolean
          example: true

    # --- Agent ---
...
    MemorySyncRequest:
      type: object
      required:
        - agentId
      properties:
        agentId:
          type: string
          format: uuid
          description: Agent whose memory files should be synced.
        direction:
          type: string
          enum:
            - file_to_qdrant
            - qdrant_to_file
            - bidirectional
          default: bidirectional
          description: |
            Sync direction. `file_to_qdrant` is the most common path
            (markdown is authoritative).  `bidirectional` performs both
            directions, with file content winning on conflicts.
        dryRun:
          type: boolean
          default: false
          description: If true, report changes without applying them.
...
  parameters:
    agentId:
      name: agentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique agent identifier.
      example: 550e8400-e29b-41d4-a716-446655440000
    jobId:
      name: jobId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Unique job identifier.
      example: 660e8400-e29b-41d4-a716-446655440001
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Maximum number of items to return.
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip.

  # -----------------------------------------------------------------------
  # Reusable responses
  # -----------------------------------------------------------------------
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication credentials.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://cortex-plane.dev/errors/unauthorized
            title: Unauthorized
            status: 401
            detail: Missing or invalid Authorization header.
...
    NotFound:
      description: The requested resource does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://cortex-plane.dev/errors/not-found
            title: Not Found
            status: 404
            detail: Agent 550e8400-e29b-41d4-a716-446655440000 is not managed by this control plane.
...
    TooManyRequests:
      description: Rate limit exceeded.
      headers:
        Retry-After:
          $ref: "#/components/headers/Retry-After"
        X-RateLimit-Limit:
          $ref: "#/components/headers/X-RateLimit-Limit"
        X-RateLimit-Remaining:
          $ref: "#/components/headers/X-RateLimit-Remaining"
        X-RateLimit-Reset:
          $ref: "#/components/headers/X-RateLimit-Reset"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetail"
          example:
            type: https://cortex-plane.dev/errors/rate-limit-exceeded
            title: Too Many Requests
            status: 429
            detail: Rate limit exceeded. Retry after 12 seconds.

# ===========================================================================
# Paths
# ===========================================================================
paths:
...
  /jobs:
    get:
      operationId: listJobs
      summary: List jobs across all agents
      tags:
        - Jobs
      parameters:
        - name: agentId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum:
              - PENDING
              - SCHEDULED
              - RUNNING
              - WAITING_FOR_APPROVAL
              - COMPLETED
              - FAILED
              - TIMED_OUT
              - RETRYING
              - DEAD_LETTER
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Paginated list of jobs for dashboard views.
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobs
                  - pagination
                properties:
                  jobs:
                    type: array
                    items:
                      type: object
                      required:
                        - id
                        - agentId
                        - status
                        - type
                        - createdAt
                      properties:
                        id:
                          type: string
                        agentId:
                          type: string
                        status:
                          type: string
                        type:
                          type: string
                        createdAt:
                          type: string
                          format: date-time
                        updatedAt:
                          type: string
                          format: date-time
                        completedAt:
                          type: string
                          format: date-time
                        error:
                          type: string
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /jobs/{jobId}:
    get:
      operationId: getJob
      summary: Get job details for drawer view
      tags:
        - Jobs
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "200":
          description: Job detail response.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - agentId
                  - status
                  - type
                  - createdAt
                  - steps
                  - logs
                properties:
                  id:
                    type: string
                  agentId:
                    type: string
                  status:
                    type: string
                  type:
                    type: string
                  createdAt:
                    type: string
                    format: date-time
                  updatedAt:
                    type: string
                    format: date-time
                  completedAt:
                    type: string
                    format: date-time
                  error:
                    type: string
                  agentName:
                    type: string
                  durationMs:
                    type: number
                  steps:
                    type: array
                    items:
                      type: object
                  logs:
                    type: array
                    items:
                      type: object
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /jobs/{jobId}/retry:
    post:
      operationId: retryJob
      summary: Retry a failed job
      tags:
        - Jobs
      parameters:
        - $ref: "#/components/parameters/jobId"
      responses:
        "202":
          description: Retry accepted.
          content:
            application/json:
              schema:
                type: object
                required:
                  - jobId
                  - status
                properties:
                  jobId:
                    type: string
                  status:
                    type: string
                    enum:
                      - retrying
        "404":
          $ref: "#/components/responses/NotFound"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # -------------------------------------------------------------------------
  # Content Pipeline
  # -------------------------------------------------------------------------
  /content:
    get:
      operationId: listContent
      summary: List content pieces for publishing pipeline
      tags:
        - Agents
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum:
              - DRAFT
              - IN_REVIEW
              - QUEUED
              - PUBLISHED
        - name: type
          in: query
          schema:
            type: string
            enum:
              - blog
              - social
              - newsletter
              - report
        - name: agentId
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Content list response.
          content:
            application/json:
              schema:
                type: object
                required:
                  - content
                  - pagination
                properties:
                  content:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/Pagination"
  /content/{id}/publish:
    post:
      operationId: publishContent
      summary: Publish content piece to channel
      tags:
        - Agents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                channel:
                  type: string
      responses:
        "200":
          description: Publish result.
          content:
            application/json:
              schema:
                type: object
                required:
                  - contentId
                  - status
                  - publishedAt
                properties:
                  contentId:
                    type: string
                  status:
                    type: string
                    enum:
                      - published
                  publishedAt:
                    type: string
                    format: date-time
  /content/{id}/archive:
    post:
      operationId: archiveContent
      summary: Archive content piece
      tags:
        - Agents
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archive result.
          content:
            application/json:
              schema:
                type: object
                required:
                  - contentId
                  - status
                properties:
                  contentId:
                    type: string
                  status:
                    type: string
                    enum:
                      - archived

  # -------------------------------------------------------------------------
  # Memory
  # -------------------------------------------------------------------------
  /memory/search:
    get:
      operationId: searchMemory
      summary: Search memory records
      tags:
        - Memory
      parameters:
        - name: agentId
          in: query
          required: true
          schema:
            type: string
        - name: query
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Memory search results.
          content:
            application/json:
              schema:
                type: object
                required:
                  - results
                properties:
                  results:
                    type: array
                    items:
                      type: object
  /memory/sync:
    post:
      operationId: syncMemory
      summary: Trigger bidirectional Markdown ↔ Qdrant sync
      description: |
        Manually triggers a memory synchronization between the agent's
        markdown files (MEMORY.md) and Qdrant vector storage.

        By default the sync is bidirectional with the markdown file as
        authoritative on conflicts ("human wins" policy).  Use the
        `direction` parameter to limit to one direction.

        Sync uses structure-aware chunking (parsing `##` headers),
        SHA-256 content hashing for deduplication, and UUIDv5
        deterministic IDs for idempotent upserts.
      tags:
        - Memory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemorySyncRequest"
            example:
              agentId: 550e8400-e29b-41d4-a716-446655440000
              direction: file_to_qdrant
              dryRun: false
      responses:
        "200":
          description: Sync completed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemorySyncResponse"
              example:
                syncId: aa0e8400-e29b-41d4-a716-446655440005
                agentId: 550e8400-e29b-41d4-a716-446655440000
                status: completed
                stats:
                  upserted: 12
                  deleted: 3
                  unchanged: 45
                  durationMs: 1823
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Agent not found or has no memory files.
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/ProblemDetail"
        "429":
          $ref: "#/components/responses/TooManyRequests"

  # -------------------------------------------------------------------------
  # Browser Observation (Dashboard)
  # -------------------------------------------------------------------------
  /agents/{agentId}/browser:
    get:
      operationId: getAgentBrowser
      summary: Get browser session info
      tags:
        - Agents
      parameters:
        - $ref: "#/components/parameters/agentId"
      responses:
        "200":
          description: Browser session info.
          content:
            application/json:
              schema:
                type: object
                required:
                  - id
                  - agentId
                  - vncUrl
                  - status
                  - tabs
                  - latencyMs
                properties:
                  id:
                    type: string
                  agentId:
                    type: string
                  vncUrl:
                    type: string
                    nullable: true
                  status:
                    type: string
                    enum:
                      - connecting
                      - connected
                      - disconnected
                      - error
                  tabs:
                    type: array
                    items:
                      type: object
                  latencyMs:
                    type: number
                  lastHeartbeat:
                    type: string
                    format: date-time
  /agents/{agentId}/browser/screenshots:
    get:
      operationId: listAgentScreenshots
      summary: List browser screenshots for an agent
      tags:
        - Agents
      parameters:
        - $ref: "#/components/parameters/agentId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: Screenshot list response.
          content:
            application/json:
              schema:
                type: object
                required:
                  - screenshots
                properties:
                  screenshots:
                    type: array
                    items:
                      type: object
  /agents/{agentId}/browser/events:
    get:
      operationId: listAgentBrowserEvents
      summary: List browser event history for an agent
      tags:
        - Agents
      parameters:
        - $ref: "#/components/parameters/agentId"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
        - name: types
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Browser event list response.
          content:
            application/json:
              schema:
                type: object
                required:
                  - events
                properties:
                  events:
                    type: array
                    items:
                      type: object

  # -------------------------------------------------------------------------
  # SSE
  # -------------------------------------------------------------------------
...